+++ Evergreen/Apps/Get-Zoom.ps1
@@ -27,11 +27,17 @@ Function Get-Zoom {
             $redirectUrl = Resolve-SystemNetWebRequest -Uri $res.Get.Download[$platform][$installer]
 
             # Match the URL without the text after the ?
-            try {
-                $Url = [RegEx]::Match($redirectUrl.ResponseUri.AbsoluteUri, $res.Get.MatchUrl).Captures.Groups[1].Value
+            If ($Null -eq $redirectUrl) {
+                Write-Verbose -Message "$($MyInvocation.MyCommand): Setting fallback URL to: $($script:resourceStrings.Uri.Issues)."
+                $Url = $script:resourceStrings.Uri.Issues
             }
-            catch {
-                $Url = $redirectUrl.ResponseUri.AbsoluteUri
+            Else {
+                try {
+                    $Url = [RegEx]::Match($redirectUrl.ResponseUri.AbsoluteUri, $res.Get.MatchUrl).Captures.Groups[1].Value
+                }
+                catch {
+                    $Url = $redirectUrl.ResponseUri.AbsoluteUri
+                }
             }
 
             # Match version number from the download URL